Add a guard so that there aren't spec failures if specs are ran outside of a git repo.

+++ spec/git-index_spec.cr
@@ -1,7 +1,10 @@
 require "./spec_helper"
 
+IN_GIT_REPO = !`git -C . config --get remote.origin.url`.strip.empty?
+
 describe GitIndex do
   it "works" do
+    if IN_GIT_REPO
     ARGV.replace ["--database",".git-index.db","-i","."]
     gi = GitIndex.new
     db = gi.database
@@ -38,7 +41,10 @@ describe GitIndex do
     processed5 = gi.list_records(db)
     pp processed5
     processed5.empty?.should be_true
+    else
+      true.should be_true
+    end
   ensure
-    File.delete(".git-index.db")
+    File.delete(".git-index.db") if File.exists?(".git-index.db")
   end
 end
